refactor: add handler
