+++ server.go
@@ -2,27 +2,13 @@ package main
 
 import (
 	"database/sql"
-	"errors"
+	"echo-me/presentation"
 	"fmt"
 	"log/slog"
-	"net/http"
 
 	"github.com/getsentry/sentry-go"
-	sentryecho "github.com/getsentry/sentry-go/echo"
-	"github.com/labstack/echo/v4"
 	_ "github.com/mattn/go-sqlite3"
 )
-
-func saveSlot(c echo.Context) error {
-	id := c.Param("id")
-	name := c.FormValue("name")
-	email := c.FormValue("email")
-	return c.String(http.StatusOK, id+". name:"+name+", email:"+email)
-}
-
-func getSlots(c echo.Context) error {
-	return c.String(http.StatusOK, "get slots")
-}
 
 func main() {
 	db, err := sql.Open("sqlite3", "./foo.db")
@@ -42,14 +28,6 @@ func main() {
 		fmt.Printf("Sentry initialization failed: %v", err)
 	}
 
-	e := echo.New()
-	e.Use(sentryecho.New(sentryecho.Options{Repanic: true}))
-	e.PUT("/slots/:id", saveSlot)
-	e.GET("/slots", getSlots)
-	e.GET("/er", getError)
-	e.Logger.Fatal(e.Start(":1323"))
-}
-
-func getError(c echo.Context) error {
-	return errors.New("mock error")
+	server := presentation.NewServer()
+	server.Start(":1323")
 }
